changed error handling to allow negative tests
process_int.h/cc:
* changed call of checkStreamCnt() to validate()

process.cc/h:
* added multiple functions
** validate - calls checkStreamCnt and checkInStreamFiles() and handles errorCodes
** handleProcessErr - handles error codes
** added enum class for error codes
* changed checkStramCnt to return errors codes instead of aborting

Comment:
Those changes will make debugging a little harder since the error message is not printed where the
error happened but is caught later. The positive aspect of this is that we can now write negative test
for the process creation since we can now call the functions and check for the codes instead of
the old behaviour where the program would exit. By looking up the error code we then can find
where the error actually happened.